moved tabs to widgets
