+++ lib/screens/animal_details_screen.dart
@@ -288,7 +288,7 @@ class AnimalDetailScreen extends ConsumerWidget {
                     fontWeight: FontWeight.bold,
                   ),
                 ),
-                SizedBox(
+                const SizedBox(
                   height: 10,
                 ),
                 Text(
