+++ lib/screens/login_screen.dart
@@ -1,6 +1,6 @@
 import 'package:adoption_app/screens/forgot_password_screen.dart';
 import 'package:adoption_app/screens/sign_up_screen.dart';
-import 'package:adoption_app/screens/tabs.dart';
+import 'package:adoption_app/widgets/tabs.dart';
 import 'package:sign_in_button/sign_in_button.dart';
 import 'package:flutter/material.dart';
 import 'package:adoption_app/services/firebase_authentication.dart';
